Grid
Implemented FlowFieldAI, CommonAI, Events in Common. Made 1st step of Grid plus gridoverlay, assigned comma as keybind to open overlay. Added events in Common to be able to throw an event to listen to from a module, so core doesn't need to know the modules that need keybinds. Put the already existing files in common into a common folder for better structure.